[Refactor] 변수명 선언 파일 위치 수정
