--- js/utils/commonsBase.js
@@ -1,17 +1,0 @@
-// signin에서 사용하는 변수
-export const signinEmailInput = querySelector("#signin-email-input");
-export const signinPasswordInput = querySelector("#signin-password-input");
-export const loginBtn = querySelector("#login-button");
-export const loginForm = querySelector("#login-form");
-export const passwordEyeIcon = querySelector("#password-eye-icon");
-
-// signup에서 사용하는 변수
-export const signupEmailInput = querySelector("#signup-email-input");
-export const signupPasswordInput = querySelector("#signup-password-input");
-export const signupPasswordRepeatInput = querySelector(
-  "#signup-password-repeat-input"
-);
-export const signupBtn = querySelector("#signup-button");
-export const signupForm = querySelector("#signup-form");
-export const eyeIconPassword = querySelector("#password-eye-icon");
-export const passwordRepeatEyeIcon = querySelector("#password-repeat-eye-icon");
